# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      parameters:
        - name: stage
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update a deal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete a deal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id: { type: integer }
        companyId: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        status: { type: string }
        source: { type: string }
        value: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, companyId, name, status, source, createdAt, updatedAt]
    LeadInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        company: { type: string }
        status: { type: string }
        source: { type: string }
        value: { type: number }
        notes: { type: string }
      required: [name]
    LeadUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        company: { type: string }
        status: { type: string }
        source: { type: string }
        value: { type: number }
        notes: { type: string }
    Deal:
      type: object
      properties:
        id: { type: integer }
        companyId: { type: integer }
        title: { type: string }
        stage: { type: string }
        value: { type: number }
        probability: { type: ["integer", "null"] }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        expectedCloseDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, companyId, title, stage, value, createdAt, updatedAt]
    DealInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        stage: { type: string }
        value: { type: number }
        probability: { type: integer }
        customerId: { type: integer }
        expectedCloseDate: { type: string }
        notes: { type: string }
      required: [title, value]
    DealUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        stage: { type: string }
        value: { type: number }
        probability: { type: integer }
        customerId: { type: integer }
        expectedCloseDate: { type: string }
        notes: { type: string }
...